cars details working

+++ Models/Car.cs
@@ -3,7 +3,7 @@ namespace CapStone2.Models;
 public class Car
 {
     public int Id { get; set; }
-    public decimal Year { get; set; }
+    public int Year { get; set; }
     public string Make { get; set; }
     public string Model { get; set; }
     public int ProfileId { get; set; }
